requirements.txt
edge-tts==7.0.0
customtkinter==5.2.2
python-docx==1.1.2
pygame==2.6.1
pydub==0.25.1
audioop-lts==0.2.1; python_version >= "3.13"
chardet==5.2.0
pyinstaller==6.11.1
requests==2.32.3
packaging==24.2
